chore: moved types out to root
