+++ packages/wallets/src/features/cfd/flows/MT5/AvailableMT5AccountsList/AvailableMT5AccountsList.tsx
@@ -6,7 +6,7 @@ import DerivedMT5 from '../../../../../public/images/mt5-derived.svg';
 import FinancialMT5 from '../../../../../public/images/mt5-financial.svg';
 import SwapFreeMT5 from '../../../../../public/images/mt5-swap-free.svg';
 import { JurisdictionModal, MT5PasswordModal } from '../../../modals';
-import { THooks } from '../../../types';
+import { THooks } from '../../../../../types';
 import './AvailableMT5AccountsList.scss';
 
 const marketTypeToDescriptionMapper = {
