+++ packages/wallets/src/features/cfd/modals/MT5PasswordModal/MT5PasswordModal.tsx
@@ -11,7 +11,7 @@ import { ModalWrapper, WalletButton } from '../../../../components/Base';
 import { useModal } from '../../../../components/ModalProvider';
 import MT5PasswordIcon from '../../../../public/images/ic-mt5-password.svg';
 import { CreatePassword, EnterPassword, Success } from '../../screens';
-import { TMarketTypes, TPlatforms } from '../../types';
+import { TMarketTypes, TPlatforms } from '../../../../types';
 
 type TProps = {
     marketType: TMarketTypes.SortedMT5Accounts;
