+++ packages/wallets/src/features/cfd/screens/CreatePassword/CreatePassword.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import PasswordShowIcon from '../../../../public/images/ic-password-show.svg';
 import './CreatePassword.scss';
-import { TPlatforms } from '../../types';
+import { TPlatforms } from '../../../../types';
 
 const platformToTitleMapper: Record<TPlatforms.All, string> = {
     ctrader: 'cTrader',
