+++ packages/wallets/src/features/cfd/screens/EnterPassword/EnterPassword.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import PasswordShowIcon from '../../../../public/images/ic-password-show.svg';
 import './EnterPassword.scss';
-import { TMarketTypes, TPlatforms } from '../../types';
+import { TMarketTypes, TPlatforms } from '../../../../types';
 
 type TProps = {
     isLoading?: boolean;
