+++ packages/wallets/src/features/cfd/screens/Jurisdiction/jurisdiction-contents/props.types.ts
@@ -1,4 +1,4 @@
-import { THooks } from '../../../types';
+import { THooks } from '../../../../../types';
 
 export type TClickableDescription = {
     onClick?: React.MouseEventHandler<HTMLSpanElement>;
