+++ packages/wallets/src/features/cfd/screens/Success/Success.tsx
@@ -4,7 +4,7 @@ import { WalletText } from '../../../../components';
 import { WalletGradientBackground } from '../../../../components/WalletGradientBackground';
 import { WalletMarketCurrencyIcon } from '../../../../components/WalletMarketCurrencyIcon';
 import './Success.scss';
-import { TDisplayBalance, TMarketTypes, TPlatforms } from '../../types';
+import { TDisplayBalance, TMarketTypes, TPlatforms } from '../../../../types';
 
 type TSuccessProps = {
     description: string;
